requirements.txt
cycler==0.10.0
matplotlib==1.5.3
numpy==1.11.3
olefile==0.44
Pillow==4.0.0
pymba==0.1
pyparsing==2.1.10
python-dateutil==2.6.0
pytz==2016.10
six==1.10.0
